+++ src/isal/__init__.py
@@ -39,4 +39,4 @@ __all__ = [
     "__version__"
 ]
 
-__version__ = "0.9.0"
+__version__ = "0.10.0"