+++ CHANGELOG.rst
@@ -7,6 +7,13 @@ Changelog
 .. This document is user facing. Please word the changes in such a way
 .. that users understand how the changes affect the new version.
 
+version 0.10.0
+------------------
++ Added an ``igzip_lib`` module which allows more direct access to ISA-L's
+  igzip_lib API. This allows features such as headerless compression and
+  decompression, as well as setting the memory levels manually.
++ Added more extensive documentation.
+
 version 0.9.0
 -----------------
 + Fix a bug where a AttributeError was triggered when zlib.Z_RLE or